requirements.txt
# Code Review Claude - Python Dependencies
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
jinja2>=3.1.0
